ibc tutorial updates for transfer/call separation
