+++ ibc-call/artifacts/checksums.txt
@@ -1,0 +1,1 @@
+49d7c198384d5f483d1a7047ffb60221a979e91a39c8475df6733e6ed05d05d1  ibc_tutorial.wasm
