+++ ibc-call/artifacts/checksums_intermediate.txt
@@ -1,0 +1,1 @@
+955302af83e80db2daaf50e362d4072e538c09ba7f8cffc2848a91b8dc8730d9  /target/wasm32-unknown-unknown/release/ibc_tutorial.wasm
