+++ ibc-call/src/lib.rs
@@ -1,0 +1,10 @@
+pub mod ack;
+pub mod contract;
+mod error;
+pub mod ibc;
+pub mod msg;
+pub mod state;
+
+
+
+pub use crate::error::ContractError;
