+++ ibc-transfer/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "ibc-tutorial"
+version = "0.1.0"
+authors = ["Coreum"]
+edition = "2021"
+
+exclude = [
+    "ibc_tutorial.wasm",
+    "checksums.txt",
+]
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[profile.release]
+opt-level = 3
+debug = false
+rpath = false
+lto = true
+debug-assertions = false
+codegen-units = 1
+panic = 'abort'
+incremental = false
+overflow-checks = true
+
+[features]
+backtraces = ["cosmwasm-std/backtraces"]
+library = []
+
+[dependencies]
+cosmwasm-std = { version = "1.3.1", features = ["ibc3"] }
+cosmwasm-storage = "1.3.1"
+cw-storage-plus = "1.1.0"
+cw2 = "1.1.0"
+schemars = "0.8.8"
+thiserror = { version = "1.0.31" }
+cosmwasm-schema = "1.3.1"
