+++ ibc-transfer/src/contract.rs
@@ -12,8 +12,7 @@ use cosmwasm_std::{
 use cw2::set_contract_version;
 
 // Local module imports for messages and state.
-use crate::msg::{ExecuteMsg, InstantiateMsg, IbcExecuteMsg, QueryMsg, GetCountResponse};
-use crate::state::{CONNECTION_COUNTS, TIMEOUT_COUNTS};  // Note: Removed TIMEOUT_COUNTS as it's not in use.
+use crate::msg::{ExecuteMsg, InstantiateMsg};
 
 // Constants for contract name and version, fetched from Cargo.toml.
 const CONTRACT_NAME: &str = env!("CARGO_PKG_NAME");
@@ -48,17 +47,6 @@ pub fn execute(
             amount,
             timeout,
         } => transfer(channel_id, to_address, amount, timeout),
-        ExecuteMsg::Increment { channel, timeout } => Ok(Response::new()
-            .add_attribute("method", "execute_increment")
-            .add_attribute("channel", channel.clone())
-            // Send an IBC message after the packet is received on the other chain.
-            .add_message(IbcMsg::SendPacket {
-                channel_id: channel,
-                data: to_binary(&IbcExecuteMsg::Increment {})?,
-                // Set a default timeout of two minutes.
-                // timeout: IbcTimeout::with_timestamp(_env.block.time.plus_seconds(120)),
-                timeout: timeout,
-            })),
     }
 }
 
@@ -80,36 +68,6 @@ fn transfer(
         .add_attribute("method", "transfer")
         .add_message(ibc_transfer_msg);
     Ok(res)
-}
-
-// Function to be called when an IBC packet is received on the other chain.
-pub fn try_increment(deps: DepsMut, channel: String) -> StdResult<u32> {
-    CONNECTION_COUNTS.update(deps.storage, channel, |count| -> StdResult<_> {
-        Ok(count.unwrap_or_default() + 1)
-    })
-}
-
-
-#[cfg_attr(not(feature = "library"), entry_point)]
-pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> StdResult<Binary> {
-    match msg {
-        QueryMsg::GetCount { channel } => to_binary(&query_count(deps, channel)?),
-        QueryMsg::GetTimeoutCount { channel } => to_binary(&query_timeout_count(deps, channel)?),
-    }
-}
-
-fn query_count(deps: Deps, channel: String) -> StdResult<GetCountResponse> {
-    let count = CONNECTION_COUNTS
-        .may_load(deps.storage, channel)?
-        .unwrap_or_default();
-    Ok(GetCountResponse { count })
-}
-
-fn query_timeout_count(deps: Deps, channel: String) -> StdResult<GetCountResponse> {
-    let count = TIMEOUT_COUNTS
-        .may_load(deps.storage, channel)?
-        .unwrap_or_default();
-    Ok(GetCountResponse { count })
 }
 
 // Unit tests for the contract.
@@ -162,26 +120,4 @@ mod tests {
         }
     }
 
-    // Test the increment functionality over IBC.
-    #[test]
-    fn increment() {
-        let mut deps = mock_dependencies();
-
-        let channel: String = "channel-0".into();  // Type annotation for clarity.
-
-        let msg = ExecuteMsg::Increment { 
-            channel: channel.clone(),
-            timeout: IbcTimeout::with_timestamp(mock_env().block.time.plus_seconds(120)),
-        };
-
-        let info = mock_info("sender", &[]);
-        let res = execute(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-        // Verify the increment message attributes.
-        assert_eq!(res.attributes[0].key, "method");
-        assert_eq!(res.attributes[0].value, "execute_increment");
-
-        assert_eq!(res.attributes[1].key, "channel");
-        assert_eq!(res.attributes[1].value, channel);
-    }
 }
