+++ ibc-transfer/src/lib.rs
@@ -1,0 +1,7 @@
+pub mod contract;
+mod error;
+pub mod msg;
+
+
+
+pub use crate::error::ContractError;
