--- src/lib.rs
@@ -1,16 +1,0 @@
-// #![recursion_limit = "1024"]
-// #![cfg_attr(feature = "nightly", feature(const_fn, unboxed_closures, abi_thiscall, const_fn_trait_bound))]
-// #![cfg_attr(
-//   all(feature = "nightly", test),
-//   feature(naked_functions, core_intrinsics, llvm_asm)
-// )]
-pub mod ack;
-pub mod contract;
-mod error;
-pub mod ibc;
-pub mod msg;
-pub mod state;
-
-
-
-pub use crate::error::ContractError;
